edmcp-editcheck/pyproject.toml
[project]
name = "edmcp-editcheck"
version = "0.1.0"
description = "FastMCP server for auditing Google Doc edit histories to detect AI/copy-paste cheating"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.0.0",
    "openai>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
